# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для конструктора email-рассылок",
        "title": "Email Campaign API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/templates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Получить все не архивированные шаблоны",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAllTemplatesResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Создать шаблон",
                "parameters": [
                    {
                        "description": "Данные шаблона",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTemplateResponse"
                        }
                    }
                }
            }
        },
...
        "/api/templates/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Получить шаблон по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetTemplateByIDResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Обновить шаблон",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные шаблона",
                        "name": "template",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/templates/{id}/archive": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Архивировать шаблон",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ArchiveTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/templates/{id}/restore": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Восстановить шаблон из архива",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.RestoreTemplateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ArchiveTemplateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "шаблон архивирован"
                },
                "template": {
                    "$ref": "#/definitions/dto.TemplateResponse"
                }
            }
        },
...
        "dto.CreateTemplateRequest": {
            "type": "object",
            "required": [
                "html_body",
                "name"
            ],
            "properties": {
                "html_body": {
                    "type": "string",
                    "example": "\u003ch1\u003eHello\u003c/h1\u003e"
                },
                "name": {
                    "type": "string",
                    "example": "Welcome template"
                }
            }
        },
        "dto.CreateTemplateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "шаблон создан"
                },
                "template": {
                    "$ref": "#/definitions/dto.TemplateResponse"
                }
            }
        },
...
        "dto.GetAllTemplatesResponse": {
            "type": "object",
            "properties": {
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TemplateResponse"
                    }
                }
            }
        },
        "dto.GetTemplateByIDResponse": {
            "type": "object",
            "properties": {
                "template": {
                    "$ref": "#/definitions/dto.TemplateResponse"
                }
            }
        },
...
        "dto.RestoreTemplateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "шаблон восстановлен"
                },
                "template": {
                    "$ref": "#/definitions/dto.TemplateResponse"
                }
            }
        },
        "dto.TemplateResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-06-01T12:00:00Z"
                },
                "created_by": {
                    "type": "integer",
                    "example": 18
                },
                "deleted_at": {
                    "type": "string",
                    "example": "2024-06-01T12:00:00Z"
                },
                "deleted_by": {
                    "type": "integer",
                    "example": 2
                },
                "html_body": {
                    "type": "string",
                    "example": "\u003ch1\u003eHello\u003c/h1\u003e"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "type": "string",
                    "example": "Welcome template"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-06-01T12:00:00Z"
                },
                "updated_by": {
                    "type": "integer",
                    "example": 18
                }
            }
        },
...
        "dto.UpdateTemplateRequest": {
            "type": "object",
            "properties": {
                "html_body": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateTemplateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "шаблон обновлён"
                },
                "template": {
                    "$ref": "#/definitions/dto.TemplateResponse"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}